Fixed buffer overflow, flipped date and time
Fixed a buffer overflow in which there was no space for the null character for a date in September. Switched date and time so that the time is visible during a toast notification. Made .h access use src/c at the top to get things to compile, which mysteriously seems required even though it wasn't before.